# api-contract/src/main/openapi/justvotes-v1.yaml
openapi: 3.0.3
info:
  title: JustVotes API
  version: 1.0.0
  description: |
    The authoritative HTTP contract for JustVotes. Poll audit entries are intentionally public and detailed so every participant can verify voting activity; they contain the public identity and selected option.

    ## Cookie contract

    The browser client uses same-origin credentials (`credentials: "same-origin"`), so it sends and stores these cookies automatically. `userID` is the normalized identity value, valid for ten years, `Path=/`, `HttpOnly`, `SameSite=Lax`, and `Secure` when the effective external request protocol is HTTPS. JavaScript must not read or mirror it. `JSESSIONID` is a server-managed session cookie: `Path=/`, session lifetime, `HttpOnly`, `SameSite=Lax`, and `Secure` for effective external HTTPS. `XSRF-TOKEN` is a CSRF bootstrap cookie: `Path=/`, session lifetime, readable by JavaScript (`HttpOnly` absent/false), `SameSite=Lax`, and `Secure` for effective external HTTPS; the client reads it only to send the matching `X-XSRF-TOKEN` header. The application trusts forwarded protocol headers only after deployment behind its trusted proxy; direct untrusted forwarding headers must not be accepted.

    ## Session and CSRF flow

    1. `GET /csrf` first. It returns `{"token":"...","headerName":"X-XSRF-TOKEN"}` and sets the readable `XSRF-TOKEN` cookie. Keep that cookie and send the JSON token in the returned header name for every unsafe request below.
...

...
servers:
  - url: /api/v1
...
paths:
...
  /admin/session:
    get:
      tags:
        - Session
      operationId: session
      security:
        - adminSession: [ ]
      responses:
        '204':
          description: Active session
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '401':
          $ref: '#/components/responses/Problem'
...
  /admin/template-catalog/templates:
    get:
      tags:
        - Templates
      operationId: templates
      security:
        - adminSession: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '200':
          description: Templates
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
    post:
      tags:
        - Templates
      operationId: createTemplate
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      requestBody:
        $ref: '#/components/requestBodies/Name'
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '201':
          description: Created
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/template-catalog/templates/{templateId}':
    parameters:
      - $ref: '#/components/parameters/TemplateId'
    patch:
      tags:
        - Templates
      operationId: renameTemplate
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      requestBody:
        $ref: '#/components/requestBodies/Name'
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Updated
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
    delete:
      tags:
        - Templates
      operationId: deleteTemplate
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '204':
          description: Deleted
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /admin/template-catalog/groups:
    get:
      tags:
        - Templates
      operationId: groups
      security:
        - adminSession: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '200':
          description: Groups
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateGroup'
    post:
      tags:
        - Templates
      operationId: createGroup
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      requestBody:
        $ref: '#/components/requestBodies/Group'
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '201':
          description: Created
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateGroup'
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/template-catalog/groups/{groupId}':
    parameters:
      - $ref: '#/components/parameters/GroupId'
    patch:
      tags:
        - Templates
      operationId: renameGroup
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      requestBody:
        $ref: '#/components/requestBodies/Name'
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Updated
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateGroup'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
    delete:
      tags:
        - Templates
      operationId: deleteGroup
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '204':
          description: Deleted
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/template-catalog/groups/{groupId}/templates':
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags:
        - Templates
      operationId: templatesInGroup
      security:
        - adminSession: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '200':
          description: Templates
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
        '404':
          $ref: '#/components/responses/Problem'
  '/admin/template-catalog/groups/{groupId}/templates/{templateId}':
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/TemplateId'
    put:
      tags:
        - Templates
      operationId: assignTemplate
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '204':
          description: Assigned
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
    delete:
      tags:
        - Templates
      operationId: removeTemplate
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '204':
          description: Removed
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Server-managed session cookie; Path=/, session lifetime, SameSite=Lax, HttpOnly=true, Secure for effective HTTPS. Sent automatically as same-origin credentials.
    voterIdentity:
      type: apiKey
      in: cookie
      name: userID
      description: Normalized identity cookie; ten-year lifetime, Path=/, SameSite=Lax, HttpOnly=true, Secure for effective HTTPS. Browser-managed and never readable by JavaScript.
    csrf:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: Send the token returned by `GET /csrf` in this header together with its `XSRF-TOKEN` cookie.
    csrfCookie:
      type: apiKey
      in: cookie
      name: XSRF-TOKEN
      description: Path=/, session lifetime, SameSite=Lax, HttpOnly=false, Secure for effective HTTPS. JavaScript reads this cookie and sends the matching token in the X-XSRF-TOKEN header.
  parameters:
...
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/GroupId'
    TemplateId:
      name: templateId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/TemplateId'
  requestBodies:
...
    Name:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Name'
    Group:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GroupInput'
...
  responses:
    Problem:
      description: Problem Details (RFC 9457). `code` is stable and suitable for machine handling; `detail` is diagnostic text.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  headers:
    NoStore:
      description: Response must not be stored or reused by browsers or intermediary caches.
      schema:
        type: string
        enum: [no-store]
  schemas:
...
    GroupId:
      type: string
      pattern: '^g_v1_[A-Za-z0-9_-]+$'
    TemplateId:
      type: string
      pattern: '^t_v1_[A-Za-z0-9_-]+$'
...
    Name:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          pattern: '.*\S.*'
    GroupInput:
      type: object
      additionalProperties: false
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          maxLength: 1_000
...
    Template:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          $ref: '#/components/schemas/TemplateId'
        name:
          type: string
    TemplateGroup:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - description
      properties:
        id:
          $ref: '#/components/schemas/GroupId'
        name:
          type: string
        description:
          type: string
...
    Problem:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string